refactor(Major): refactor most code, fixing the sidebar
