+++ .gitignore
@@ -1,0 +1,3 @@
+# Ignore Python Cache Folder
+__pycache__/
+blueprints/__pycache__/
