+++ templates/user/dashboard.html
@@ -14,7 +14,7 @@
           <!-- small box -->
           <div class="small-box bg-aqua">
             <div class="inner">
-              <h3>150</h3>
+              <h3>150 </h3>
 
               <p>New Orders</p>
             </div>
